redux was added and the creation of the course was unified (#47)
* empezando

* para empezar cambio unificacion

* v1 create course form
