+++ src/App.js
@@ -17,12 +17,6 @@ import InstructorPerformanceOverview from './pages/InstructorPerformanceOverview
 import InstructorIndex from './pages/Instructor.index';
 import PerformanceIndex from './pages/Peformance.index';
 import CoursesManage from './pages/CoursesManage';
-import CoursesManageGoals from './pages/CoursesManageGoals';
-import CoursesManageCurriculum from './pages/CoursesManageCurriculum';
-import ManageIndex from './pages/ManageIndex';
-import CoursesManagePricing from './pages/CoursesManagePricing';
-import CoursesManageCourselLandingPage from './pages/CoursesManageCourselLandingPage';
-
 import './App.css';
 import Course from './pages/Course/Course';
 import Cart from './pages/Cart';
@@ -53,13 +47,8 @@ function App() {
             </Route>
           </Route>
 
-          <Route  path="instructor/courses/:course/manage" element={<CoursesManage />}>
-              <Route index element={<ManageIndex />}/>
-              <Route path="goals" element={<CoursesManageGoals/>} />
-              <Route path="curriculum" element={<CoursesManageCurriculum />} />
-              <Route path="basics" element={<CoursesManageCourselLandingPage />} />
-              <Route path="pricing" element={<CoursesManagePricing />} />
-            </Route>
+          <Route  path="instructor/courses/:course/manage" element={<CoursesManage />} />
+
           <Route path="/course/pythonforbeginners" element={<Course />} />
           <Route path="/cart" element={<Cart />} />
           <Route path='/test-cloudify' element={<TestCloudify/>}/>
