--- src/pages/CoursesManageCurriculum.jsx
@@ -1,14 +1,0 @@
-import TitleManageCourse from "../components/TitleManageCourse";
-import CoursesManageNav from '../components/CoursesManageNav';
-function CoursesManageCurriculum(){
-    return(
-        <div className="main-container-manage-goals">
-            <CoursesManageNav/>
-            <TitleManageCourse
-            title='Curriculum' />
-            <div className="container-manage-goals">
-            </div>
-        </div>
-    )
-}
-export default CoursesManageCurriculum;
