+++ src/store/InstructorCourses.Slice.js
@@ -1,20 +1,26 @@
 import { createSlice } from '@reduxjs/toolkit';
 
+
 const initialState = {
-    courses:[]
+    courses:[],
+    petition:false
 }
 
 export const intructorCourses = createSlice({
   name: 'InstructorCourses',
   initialState,
   reducers: {
-    SetCourses: (state,action) => {
-      state.courses.push(action.payload);
+    CreateCourse:(state,action)=>{
+      console.log(action.payload)
+    },
+    SetPetition:(state,action)=>{
+      state.petition = action.payload
     }
+
   }
 })
 
 
-export const { SetCourses } = intructorCourses.actions
+export const { CreateCourse , SetPetition } = intructorCourses.actions
 
 export default intructorCourses.reducer
