+++ src/styles/pages/CoursesManageGoals.scss
@@ -4,6 +4,7 @@
     width: 100%;
     border: 1px solid #d1d7dc;
     max-width: 1030px;
+    margin: 0 auto;
     box-shadow: 0 2px 4px rgb(0 0 0 / 8%), 0 4px 12px rgb(0 0 0 / 16%);
     .container-manage-goals{
         padding: 48px;
